Improve the browser queries
Now that reportupdater is running in stat1002 with the hdfs user,
the scripts can use the hive command and drop the hardcoded user
and host.

In addition to that, this change set adds 4 more scripts that
query for simpler datasets that should be used to populate
the simpler first version of the browser breakdown dashboard.

To format those queries properly, we need a pivoting script
that dynamically transforms the data from hive to the expected
dashiki format. This pivoting can not be done in hql because
the set of values of the pivoting column is unknown, and hive
does not support prepared statements.

Bug: T127326
Change-Id: Ia47e2aa48eb3a573daea7c2f5cc46f6e33e7d671