Update to include MW 1.35 and 1.36 and PHP 7.3 and 7.4.
New versions need to be added before they begin to be used or else
they show up as Other in https://pingback.wmflabs.org/#media-wiki-version
and https://pingback.wmflabs.org/#php-version/php-version-timeseries.
Once this is merged, the version reports should be re-run starting at
2018-10-24 to fix the graph for 1.33 and 1.34 (which were fixed in
https://gerrit.wikimedia.org/r/c/analytics/reportupdater-queries/+/545917),
and 1.35 (which should be fixed by this patch). This patch adds 1.36 also
for the future.

Change-Id: I1c63141dce7c68c55173ea314f9bfb33334c9db5