Adapt pingback queries to use the pingback heartbeat
The hearbeat introduced in June 2018 makes every wiki
re-send pingbacks while active (with a minimum interval
of 30 days between pingbacks).

Thus we assume that if a wiki has emmited 1+ pingbacks
during the last 30 days it is considered active.
Otherwise, it is considered inactive and it does not
show up in this report's breakdowns.

Bug: T246154
Change-Id: I9c43add1688d79cfb48f9ab8885096725e8545ce