Update pingback MediaWiki versions to include new values
https://pingback.wmflabs.org/#media-wiki-version/media-wiki-version-timeseries
only includes MediaWiki versions through MediaWiki 1.38.
MediaWiki 1.39 has been released, and MediaWiki 1.40 is in
development. The list of versions in the version_simple needs
to be updated to include those versions and a few more so we
don't need to do this again too soon. The php_drilldown buckets
also need to include the new versions.

Bug: T326825
Change-Id: Ieed859ab581a7012e9efd46dd94f45d08e9f5792